+++ packages/node-backend/setup.ts
@@ -1,4 +1,4 @@
-import { RabbitMQContainer } from '@testcontainers/rabbitmq'
+import { RedisContainer } from '@testcontainers/redis'
 import { KafkaContainer } from '@testcontainers/kafka'
 import { writeFileSync } from 'fs'
 import * as fs from 'fs'
@@ -6,38 +6,24 @@ import { resolve } from 'path'
 // This file runs before jest sets the env
 // so we need to load dotenv manually if we want to use env
 
-export const rabbit = new RabbitMQContainer(
-  'rabbitmq:3.12.11-management-alpine'
-)
-
-const kafka = await new KafkaContainer('confluentinc/cp-kafka:7.9.0')
+const redis = await new RedisContainer('redis:7.2')
 
 export default async () => {
   // Replace this with your actual async function
-  const rabbitMQContainer = await rabbit.start()
-  const kafkaContainer = await kafka.start()
+  const redisContainer = await redis.start()
 
-  const rabbitMQUrl = `amqp://${rabbitMQContainer.getHost()}:${rabbitMQContainer.getMappedPort(
-    5672
-  )}`
-
-  const kafkaUrl = `${kafkaContainer.getHost()}:${kafkaContainer.getMappedPort(
-    9093
-  )}`
-
-  console.log({ kafkaUrl })
+  const redisUrl = redisContainer.getConnectionUrl()
 
   const data = {
-    rabbitMQUrl,
-    kafkaUrl
+    redisUrl
   }
 
   const filePath = resolve(__dirname, 'tempData.json')
   writeFileSync(filePath, JSON.stringify(data), 'utf-8')
 
   return async () => {
-    await rabbitMQContainer.stop()
-    await kafkaContainer.stop()
+    await redisContainer.stop()
+
     if (fs.existsSync(filePath)) {
       fs.unlinkSync(filePath)
     }