+++ packages/node-backend/vite.config.js
@@ -2,6 +2,13 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    globalSetup: './setup.ts'
+    globalSetup: './setup.ts',
+    environment: 'node',
+    globals: true,
+    server: {
+      deps: {
+        inline: ['keyv', '@keyv/redis']
+      }
+    }
   }
 })