+++ packages/node-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.5.20
 
+## 0.0.3
+
+### Patch Changes
+
+- Add node-backend
+
 ## 0.0.2
 
 ### Patch Changes